norma

+++ libs/libft/ft_putnbr_fd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: weambros <weambros@student.42.fr>          +#+  +:+       +#+        */
+/*   By: weambros <weambros@student.21-school.ru    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/06 13:49:54 by weambros          #+#    #+#             */
-/*   Updated: 2021/04/13 00:59:57 by weambros         ###   ########.fr       */
+/*   Updated: 2021/04/16 04:26:07 by weambros         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	ft_putnbr_fd(int n, int fd)
 {
-	unsigned long	nb;
+	unsigned int	nb;
 
 	if (n < 0)
 		ft_putchar_fd('-', fd);
 	if (n < 0)
-		nb = -(n);	
+		nb = -(n);
 	else
-		nb = n;	
+		nb = n;
 	if (nb >= 10)
 	{
 		ft_putnbr_fd(nb / 10, fd);
